leaflens_backend/requirements.txt
# LeafLens Backend - Requirements
# @Maharsh Doshi

# Django
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3

# ML / AI
tensorflow>=2.12
numpy>=1.24
Pillow>=10.0

# Weather API
requests>=2.31

# Environment
python-dotenv>=1.0
